requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.42.0
django-ckeditor==5.9.0
django-crispy-forms==1.9.2
django-js-asset==1.2.2
django-taggit==1.3.0
djongo==1.3.3
idna==2.10
oauthlib==3.1.0
Pillow==7.2.0
pymongo==3.11.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
sqlparse==0.2.4
urllib3==1.25.9
